refactor(websites): extraer sub-componentes de login y SettingsPanel
- login/page.tsx (2009 → 1505 líneas): extraer 7 sub-componentes 2FA
  a _components/TwoFactorComponents.tsx + helpers a login-helpers.ts
- SettingsPanel.tsx (1594 → 1189 líneas): extraer SEO score, ImageUpload,
  SharePreview, Section accordion y constantes a settings-panel-helpers.tsx
- Limpiar imports no usados (lucide icons, next/image, ui components)

Issue #244 batch 2
