+++ frontend/src/app/(tenant)/dashboard/settings/login/_components/login-helpers.ts
@@ -1,0 +1,39 @@
+import { ApiError } from '@/lib/api/client';
+
+export async function copyToClipboard(text: string): Promise<boolean> {
+  if (typeof navigator === 'undefined' || !navigator.clipboard) return false;
+  try {
+    await navigator.clipboard.writeText(text);
+    return true;
+  } catch {
+    return false;
+  }
+}
+
+export function downloadTxt(filename: string, content: string): void {
+  if (typeof window === 'undefined') return;
+  const blob = new Blob([content], { type: 'text/plain;charset=utf-8' });
+  const url = URL.createObjectURL(blob);
+  const a = document.createElement('a');
+  a.href = url;
+  a.download = filename;
+  document.body.appendChild(a);
+  a.click();
+  document.body.removeChild(a);
+  URL.revokeObjectURL(url);
+}
+
+export function extractErrorMessage(error: unknown, fallback: string): string {
+  if (error instanceof ApiError) return error.message;
+  if (error instanceof Error) return error.message;
+  return fallback;
+}
+
+export function maskEmail(email: string): string {
+  const [local, domain] = email.split('@');
+  const masked =
+    local.length <= 2
+      ? local
+      : local.slice(0, 2) + '\u2022'.repeat(Math.min(local.length - 2, 6));
+  return `${masked}@${domain}`;
+}
